Fix warning.

+++ libyara/lexer.c
@@ -3670,7 +3670,7 @@ int yr_lex_parse_rules_fd(
   #endif
 
   yyset_extra(compiler, yyscanner);
-  yy_scan_bytes((const char*) buffer, file_size, yyscanner);
+  yy_scan_bytes((const char*) buffer, (int) file_size, yyscanner);
   yyparse(yyscanner, compiler);
   yylex_destroy(yyscanner);
